update bert embeddings using TF, add bert argument extractors
